requirements.txt
yfinance==1.1.0
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.1
python-dotenv==1.0.1
scipy==1.12.0